Introduce uses of buildTarget/inverseSources
Previously, Metals would only use the result of `buildTarget/sources` to
figure which target owns a given source file. Because
`buildTarget/sources` is issued as part of build import, the list of
sources (directories and files) returned by the build server could be
stored and re-used to immediately determine the owning target, given a
source file.

It turns out, however, that for some targets this knowledge can become
outdated. For instance, Bloop supports defining the source files as a
set of globs [1]. In this case, Bloop will respond to the
`buildTarget/sources` request issued during build import with the list
of files that match at this point in time. If the developer create a new
file which would be matched by Bloop's globs, Metals will be unable to
associate it with a target.

Determining which targets own a given source file is supported by the
BSP protocol via the `buildTarget/inverseSources` request [2].

This commit introduces a new method `BuildTargets#inverseSourcesBsp`,
which finds the target owning the given source file, just like the
existing `BuildTargets#inverseSources` does, but using BSP if the
original method couldn't find a match. Because this new method is
asynchronous (it communicates with the BSP server), migrating all
call-sites to use it would require many changes to Metals.

[1]: https://github.com/scalacenter/bloop/blob/a9b3ea59631ca300c06394effbdf4b2a8e88dbf1/config/src/main/scala/bloop/config/Config.scala#L231-L236
[2]: https://build-server-protocol.github.io/docs/specification.html#inverse-sources-request